refactor: Delimiters 와 Numbers 로 클래스 분리
